+++ vkconfig/vulkan.cpp
@@ -105,6 +105,10 @@ std::string GenerateVulkanStatus() {
         return log;
     } else {
         log += format("- Vulkan Loader version: %s\n", loader_version.str().c_str());
+        const LoaderMessageLevel loader_debug_message = configurator.environment.GetLoaderMessage();
+        if (loader_debug_message != LAODER_MESSAGE_NONE) {
+            log += format("    - VK_LOADER_DEBUG=%s\n", GetLoaderDebugToken(loader_debug_message).c_str());
+        }
     }
 
     log += GetUserDefinedLayersPathsLog("VK_LAYER_PATH environment variable", USER_DEFINED_LAYERS_PATHS_ENV);